# docs/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ActiveContext": {
        "additionalProperties": false,
        "properties": {
          "org_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "station_id"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "description": "Stable machine error code",
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "message": {
            "description": "Safe user message",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "OpenShiftRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/OpenShiftRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "backfill_approver": {
            "type": "string"
          },
          "backfill_reason": {
            "type": "string"
          },
          "backfilled": {
            "type": "boolean"
          },
          "opened_at": {
            "type": "string"
          },
          "original_event_date": {
            "type": "string"
          },
          "shift_ke": {
            "format": "int64",
            "type": "integer"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "opened_at",
          "backfilled",
          "original_event_date",
          "shift_ke",
          "backfill_approver",
          "backfill_reason"
        ],
        "type": "object"
      },
...
      "Result": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/Result.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "replay": {
            "type": "boolean"
          },
          "report_id": {
            "type": "string"
          },
          "request_hash": {
            "contentEncoding": "base64",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "replay"
        ],
        "type": "object"
      },
...
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/SessionView.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "active_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "station_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "roles",
          "org_id",
          "station_ids",
          "active_context"
        ],
        "type": "object"
      },
      "Shift": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/Shift.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "business_date": {
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "price_map_hash": {
            "contentEncoding": "base64",
            "type": "string"
          },
          "price_map_snapshot": {
            "contentEncoding": "base64",
            "type": "string"
          },
          "shift_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          },
          "station_seq": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "supervisor_id": {
            "type": "string"
          },
          "timezone_snapshot": {
            "type": "string"
          }
        },
        "required": [
          "shift_id",
          "org_id",
          "station_id",
          "station_seq",
          "supervisor_id",
          "opened_at",
          "timezone_snapshot",
          "business_date",
          "status"
        ],
        "type": "object"
      },
...
      "SubmitRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/SubmitRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "claim_token": {
            "type": "string"
          },
          "draft_id": {
            "type": "string"
          },
          "payload": {},
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "shift_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "shift_id",
          "draft_id",
          "claim_token",
          "revision",
          "payload"
        ],
        "type": "object"
      },
      "Summary": {
        "additionalProperties": false,
        "properties": {
          "business_date": {
            "type": "string"
          },
          "current_report_id": {
            "type": "string"
          },
          "opened_at": {
            "type": "string"
          },
          "shift_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          },
          "station_seq": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "supervisor_id": {
            "type": "string"
          }
        },
        "required": [
          "shift_id",
          "station_id",
          "station_seq",
          "supervisor_id",
          "opened_at",
          "business_date",
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "httpOnly session cookie",
        "in": "cookie",
        "name": "pomkita_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "PomKita API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read the current session",
        "x-permitted-roles": [
          "Operator",
          "Supervisor",
          "Station Admin",
          "Owner",
          "Superadmin"
        ]
      }
    },
...
    "/api/v1/shifts": {
      "get": {
        "operationId": "listShifts",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "station_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Summary"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List shifts"
      },
      "post": {
        "operationId": "openShift",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenShiftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Open a shift"
      }
    },
...
    "/api/v1/submissions": {
      "post": {
        "operationId": "submitReport",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Submit a report"
      }
    },
...
  },
...
}